Added support for required properties.
Added tests to ensure that generated data matches the schema for
valid values and doesn't match the schema for invalid values (not
including values like NaN that are important for JavaScript but
don't seem to be supported by JSON schema).

Supressing adding values of undefined to objects when the field is
not required. In this case "undefined" means missing and is actually
valid when the field is not required.